Add new start_containers_with_systemd property to support noble stemcells - Noble requires systemd be the entrypoint of the container. We could do something in the CPI   here to detect what the stemcell needs (maybe a bump to the stemcell api version), we are   not doing that work at this time since the warden cpi currently does not support mixed   noble/pre-noble deployments. - Also pull the config struct out of the main package so we can use it in other packages.
