+++ jobs/docker_cpi/templates/cpi.json.erb
@@ -1,6 +1,7 @@
 <%=
 
 config = {
+  "start_containers_with_systemd" => p("docker_cpi.start_containers_with_systemd"),
   "Actions" => {
     "Docker" => {
       "host"        => p("docker_cpi.docker.host"),
