--- src/bosh-docker-cpi/main/config.go
@@ -1,44 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-
-	bosherr "github.com/cloudfoundry/bosh-utils/errors"
-	boshsys "github.com/cloudfoundry/bosh-utils/system"
-
-	"bosh-docker-cpi/cpi"
-)
-
-type Config struct {
-	Actions cpi.FactoryOpts
-}
-
-func NewConfigFromPath(path string, fs boshsys.FileSystem) (Config, error) {
-	var config Config
-
-	bytes, err := fs.ReadFile(path)
-	if err != nil {
-		return config, bosherr.WrapErrorf(err, "Reading config '%s'", path)
-	}
-
-	err = json.Unmarshal(bytes, &config)
-	if err != nil {
-		return config, bosherr.WrapError(err, "Unmarshalling config")
-	}
-
-	err = config.Validate()
-	if err != nil {
-		return config, bosherr.WrapError(err, "Validating config")
-	}
-
-	return config, nil
-}
-
-func (c Config) Validate() error {
-	err := c.Actions.Validate()
-	if err != nil {
-		return bosherr.WrapError(err, "Validating Actions configuration")
-	}
-
-	return nil
-}
